feat(config-helper): implements lazy prop calculation; restructures code
