+++ packages/config-helper/src/index.ts
@@ -1,7 +1,8 @@
-export * from './lib/config-helper';
-export * from './lib/schema';
-export * from './lib/normalized-schema';
 export * from './lib/config';
 export * from './lib/config-options';
-export * from './lib/properties';
+export * from './lib/create-config';
+export * from './lib/normalized-schema';
 export * from './lib/parm';
+export * from './lib/properties';
+export * from './lib/schema';
+export * from './lib/transformer';
