+++ packages/config-helper/src/lib/utils/string-util.ts
@@ -8,5 +8,3 @@ export function prefixStringIfDefined(prefix: string, str: string | null | undef
 	}
 	return prefix + str;
 }
-
-export type MustHaveNull<T> = null extends T ? T : never;
